doc: updates README.md with new folder structure refactor: rearranges page views into src/views folder refactor: extract components into separate .js files; creates Header for TJS animation refactor: creates Landing.js view for home page items, keep Splash as a layout page
