+++ public/index.html
@@ -11,6 +11,7 @@
     content="Tech talks, Workshops and other meetups for Toronto's JavaScript enthusiasts!">
   <meta property="og:image" content="https://github.com/torontojs/torontojs.com/blob/master/public/apple-icon.png">
   <meta property="og:url" content="http://torontojs.com/">
+
   <!-- End Social Previews -->
 
   <link rel="shortcut icon" href="%PUBLIC_URL%/images/favicon.ico">
