+++ public/pages/mission.md
@@ -1,10 +1,9 @@
-# Mandate 
+# Mandate
 
-TorontoJS is a not-for-profit corporation with a mandate to support all in their learning and passion for sharing knowledge on JavaScript and software development in Toronto and GTA.
+TorontoJS is a [not-for-profit corporation](https://ised-isde.canada.ca/cc/lgcy/fdrlCrpDtls.html?corpId=10120456&V_TOKEN=null&crpNm=Toronto%20JS&crpNmbr=&bsNmbr=) with a mandate to support all in their learning and passion for sharing knowledge on JavaScript and software development in Toronto and GTA. We do that through events, workshops, and signal-boosting the efforts of Toronto JS community members. Our programming is community-led; funds raised go towards supporting community efforts. Our organizers and volunteers are software engineers that believe in the power of community to enable and elevate each other.
 
 # Mission
 
-Facilitate events, connect people, share knowledge and shepard projects in order to build an open and accessible ecosystem starting with Toronto, then beyond!
+Facilitate events, connect people, share knowledge and shepard projects in order to build an open and accessible ecosystem starting with Toronto and beyond!
 
 [Code Of Conduct](https://torontojs.com/p/code_of_conduct) is enforced at all events.
-
