+++ public/pages/volunteers.md
@@ -1,10 +1,10 @@
-## Volunteers 
+## Volunteers
 
-Volunteers and their contributions are what has kept this going, making Toronto JS the welcoming and expansive place it is today. 
+Volunteers and their contributions are what has kept this going, making Toronto JS the welcoming and expansive place it is today.
 
 <br/>
 
-Many are self-taught, some are beginning their programming journey, others more experienced developers. We all believe there can be a space for people to practice, discuss and share knowledge about code as craft, and have honest discussions about engineering practice and career. 
+Many are self-taught, some are beginning their programming journey, others more experienced developers. We all believe there can be a space for people to practice, discuss and share knowledge about code as craft, and have honest discussions about engineering practice and career.
 
 <br/>
 Everyone supports events by offering their unpaid time, connections and expertise to empower and elevate others. Whether it is event support, promotion, event organizing, mediating conflicts and removing trolls. This community would not be possible without the participation of our volunteers.
@@ -13,10 +13,25 @@ Everyone supports events by offering their unpaid time, connections and expertis
 We are grateful for the time and effort they put into making this community a welcoming place for all in tech.
 
 <br/>
+
 All volunteers agree to model and uphold the [Code Of Conduct](https://torontojs.com/p/code_of_conduct)
 
 <br/>
 <br/>
+
+## Contributors
+
+This website was created and maintained, on and off by:
+
+[Michael Svatos](https://github.com/michalsvatos)
+
+[Alex Wilmer](https://twitter.com/benevolentNinja)
+
+[Kieran Huggins](https://kieran.ca/)
+
+[Jen Chan](https://jenchan.biz)
+
+// whomever makes an update post 2019, please add your name!
 
 ## Promotion, Moderation, Community Engagement
 
@@ -60,6 +75,10 @@ localost8080
 
 [Liz McReady](https://gingerkiwi.blog/)
 
+[Divish Ram](https://www.linkedin.com/in/divish-ram-694b55230)
+
+[Rachael Concessio](https://www.linkedin.com/in/rachaelconcessio/)
+
 ## Educational Accomplices
 
 [Nicole Chung](https://twitter.com/redconservatory)
@@ -74,31 +93,23 @@ localost8080
 
 [Niya Panamdanam](https://www.linkedin.com/in/niya-panamdanam/)
 
-[Rachael Concessio](https://www.linkedin.com/in/rachaelconcessio/)
-
 ## Organizers
+
+[Yuri Yang](https://www.linkedin.com/in/07yuri/)
+
+[Aman Aalam](https://www.linkedin.com/in/amanaalam/)
+
+[Sami Xie](https://www.linkedin.com/in/sami-xie-91bb4814a)
+
+## Former Organizers
+
+TorontoJS thanks all who had a hand shaping the culture we have today!
 
 [Sherry Yang](https://5hel2l2y.github.io/)
 
 [Danny Kim](https://www.linkedin.com/in/0916dhkim/)
 
-[Yuri Yang](https://www.linkedin.com/in/07yuri/)
-
-[Jen Chan](https://jenchan.biz)
-
-## Contributors
-
-This website was created and maintained, on and off by:
-
-[Michael Svatos](https://github.com/michalsvatos)
-
-[Alex Wilmer](https://twitter.com/benevolentNinja)
-
-[Kieran Huggins](https://kieran.ca/)
-
-[Jen Chan](https://jenchan.biz)
-
-## Leadership
+[Chris Naismith](https://www.linkedin.com/in/chris-naismith/)
 
 ## Directors
 
@@ -108,5 +119,16 @@ This website was created and maintained, on and off by:
 
 [Dann Toliver](https://twitter.com/danntoliver)
 
+## Founding Organizers
+
+[Matthew Mihok](https://twitter.com/mihok)
+
+[Dann Toliver](https://twitter.com/danntoliver)
+
+[Paul Dowman](https://www.linkedin.com/in/pdowman/)
+
+[Taz Singh](https://twitter.com/tazsingh)
+
 ## Advisory
+
 [Taz Singh](https://twitter.com/tazsingh) - Founder
