+++ src/components/png/index.js
@@ -1,0 +1,2 @@
+export { default as guildLogo } from './guild_logo.png'
+export { default as logo } from './logo_circle.png'
