+++ src/views/ErrorPage.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Container } from '../components/Container'
+
+const ErrorPage = () => (
+  <div>
+    <Container>
+      <p>Uh oh! Something went wrong.</p>
+      <p>[Insert witty gif here]</p>
+    </Container>
+  </div>
+)
+
+export default ErrorPage
